refactor(agent-task): colocate settings form type
