+++ libs/domains/service-settings/feature/src/lib/agentic-workflow-settings/agentic-workflow-automations-settings/agentic-workflow-automations-settings.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react'
 import { type UseFormReturn } from 'react-hook-form'
 import { AutomationSheet } from '@qovery/domains/services/feature'
 import { Button, Icon } from '@qovery/shared/ui'
+import { type AgenticWorkflowSettingsFormValues } from '../agentic-workflow-settings'
 import { AgenticWorkflowSettingsCard } from '../agentic-workflow-settings-card'
-import { type AgenticWorkflowSettingsFormValues } from '../agentic-workflow-settings.types'
 
 export function AgenticWorkflowAutomationsSettings({
   form,
